[FLINK-20250][table-runtime] NPE when invoking AsyncLookupJoinRunner#close method
